♻️ refactor: move ProductGallery to organism directory and update imports
